refactor: 迁移网络线程到 service
